requirements.txt
darkdetect
google-generativeai
openai
keyboard
pyperclip
pywin32
PySide6
